BugFix: InputGeometry task reads the wrong file
InputGeometry read the wrong file, because the geometries were taken from a single global list of all molecules. The problem: Every molecule was asigned an index within its set (training set, validation set, training set). So the index of a molecule, didn't match the its index in the global list of molecule geometries. Got rid of stupid global list and passed the path to the xyz file as parameter to the Task.